+++ modules/dxp/apps/saml/saml-opensaml-integration/src/main/java/com/liferay/saml/opensaml/integration/internal/binding/BaseSamlBinding.java
@@ -14,8 +14,6 @@
 
 package com.liferay.saml.opensaml.integration.internal.binding;
 
-import com.liferay.saml.opensaml.integration.SamlBinding;
-
 import java.util.function.Supplier;
 
 import org.opensaml.messaging.decoder.servlet.HttpServletRequestMessageDecoder;
