+++ modules/dxp/apps/saml/saml-opensaml-integration/src/main/java/com/liferay/saml/opensaml/integration/internal/metadata/MetadataManager.java
@@ -12,7 +12,7 @@
  *
  */
 
-package com.liferay.saml.opensaml.integration.metadata;
+package com.liferay.saml.opensaml.integration.internal.metadata;
 
 import com.liferay.saml.runtime.SamlException;
