+++ modules/dxp/source-formatter-api-signature-check-exceptions.txt
@@ -13,6 +13,8 @@ com.liferay.saml.opensaml.integration.internal.metadata.MetadataGeneratorUtil#bu
 com.liferay.saml.opensaml.integration.internal.metadata.MetadataGeneratorUtil#buildIdpSsoDescriptor(HttpServletRequest,String,boolean,boolean,Credential)
 com.liferay.saml.opensaml.integration.internal.metadata.MetadataGeneratorUtil#buildSpEntityDescriptor(HttpServletRequest,String,boolean,boolean,boolean,boolean,Credential)
 com.liferay.saml.opensaml.integration.internal.metadata.MetadataGeneratorUtil#buildSpSsoDescriptor(HttpServletRequest,String,boolean,boolean,boolean,Credential)
+com.liferay.saml.opensaml.integration.internal.metadata.MetadataManager#getEntityDescriptor(HttpServletRequest)
+com.liferay.saml.opensaml.integration.internal.metadata.MetadataManager#getSecurityMessageHandler(HttpServletRequest,String,boolean)
 com.liferay.saml.opensaml.integration.internal.profile.BaseProfile#decodeSamlMessage(HttpServletRequest,HttpServletResponse,SamlBinding,boolean)
 com.liferay.saml.opensaml.integration.internal.profile.BaseProfile#getSamlMessageContext(HttpServletRequest,HttpServletResponse)
 com.liferay.saml.opensaml.integration.internal.profile.BaseProfile#getSamlMessageContext(HttpServletRequest,HttpServletResponse,String)
@@ -20,8 +22,6 @@ com.liferay.saml.opensaml.integration.internal.profile.BaseProfile#getSamlSpSess
 com.liferay.saml.opensaml.integration.internal.profile.BaseProfile#getSamlSpSessionKey(HttpServletRequest)
 com.liferay.saml.opensaml.integration.internal.profile.BaseProfile#getSamlSsoSessionId(HttpServletRequest)
 com.liferay.saml.opensaml.integration.internal.profile.BaseProfile#logout(HttpServletRequest,HttpServletResponse)
-com.liferay.saml.opensaml.integration.metadata.MetadataManager#getEntityDescriptor(HttpServletRequest)
-com.liferay.saml.opensaml.integration.metadata.MetadataManager#getSecurityMessageHandler(HttpServletRequest,String,boolean)
 com.liferay.saml.opensaml.integration.resolver.UserResolver#resolveUser(UserResolverSAMLContext,ServiceContext)
 com.liferay.saml.runtime.profile.SingleLogoutProfile#getSamlSpSession(HttpServletRequest)
 com.liferay.saml.runtime.profile.SingleLogoutProfile#isSingleLogoutSupported(HttpServletRequest)
